### 2024-11-14 #7480 @editorialbot
[PRE REVIEW]: EvoVis: Dashboard for visualizing Evolutionary Neural Architecture Search algorithms
**Submitting author:** @leadang42 (<a href="http://orcid.org/0009-0005-6056-0725">Lea Dang</a>)
**Repository:** https://github.com/ankilab/EvoVis.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
