### 2024-11-14 #7480 @crvernon
:wave: @leadang42 - I am going to run this one by our larger editorial board for review to ensure it meets our requirements for substantial scholarly effort.  I'll get back in touch ASAP.  Thank you!
