### 2024-11-14 #7480 @anki-xyz
Background to this paper: This work has been led by @leadang42 and is the result of her BSc thesis. With this tool, we are able to not only monitor the progress of evolutionary neural architecture search but also perform meta analyses on the optimization procedure.

What do we need to do with the missing DOIs? Update the paper?
