### 2024-11-14 #7480 @crvernon
@anki-xyz - concerning the DOI, you can ensure that the DOI are present in every reference in the bib file where possible.  Thank you!
