### 2024-11-15 6c8c0e0
Added doi to references
