+++ paper.bib
@@ -6,6 +6,7 @@
     archivePrefix={arXiv},
     eprint={2301.08727},
     primaryClass={cs.LG},
+    doi={10.48550/arXiv.2301.08727},
 }
 
 @article{NAS-review-efficient-performance-prediction,
@@ -16,6 +17,7 @@
     number={1},
     pages={100002},
     year={2022},
+    doi={10.1016/j.jai.2022.100002},
 }
 
 @article{EvoNAS-survey,
@@ -26,6 +28,7 @@
     volume={34},    
     number={2},
     pages={550–570},
+    doi={10.1109/TNNLS.2021.3100554},
 }
 
 @article{evolutionary-computation,
@@ -36,6 +39,7 @@
     volume={82},
     number={1},
     pages={1-6},
+    doi={10.1016/S0020-0190(02)00204-1},
 }
 
 @inproceedings{EvoNAS-survey-advacements,
@@ -44,15 +48,16 @@
     booktitle={2021 IEEE Congress on Evolutionary Computation (CEC)}, 
     pages={950-957},
     year={2021},
-    organization={IEEE}
+    organization={IEEE},
+    doi={10.1109/CEC45853.2021.9504890},
 }
 
-# Not sure if this cited correctly  (https://onlinelibrary.wiley.com/doi/abs/10.1002/9780470686652.eae495)
 @article{optimization-algorithms,
     title={Review of optimization techniques},
     author={Venter, Gerhard},
     year={2010},
     journal={Encyclopedia of aerospace engineering},
+    doi={10019.1/14646},
 }
 
 @article{evolutionary-algorithm, 
@@ -63,6 +68,7 @@
     volume={4}, 
     number={3}, 
     pages={178–195},
+    doi={10.1002/widm.1124},
 } 
 
 @inproceedings{EvoNAS-NSGA-Net-multi-obj,
@@ -71,6 +77,7 @@
     booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference},
     pages={419--427},
     year = {2019},
+    doi={10.1145/3321707.332172},
 }
 
 @inproceedings{EvoNAS-rene, 
@@ -81,7 +88,8 @@
     volume={},
     number={},
     year={2023},
-    organization={IEEE}
+    organization={IEEE},
+    doi={10.1109/COINS57856.2023.10189194},
 }
 
 @inproceedings{EvoNAS-regularized-evolution-img-classifier, 
@@ -91,7 +99,8 @@
     volume={33},
     number={01},
     pages={4780--4789},
-    year={2019}
+    year={2019},
+    doi={10.1609/aaai.v33i01.33014780},
 }
 
 @article{EvoNAS-CNN-design,
@@ -102,6 +111,7 @@
     volume={50},
     number={9},
     pages={3840-3854},  
+    doi={10.1109/TCYB.2020.2983860}
 }
 
 # Not sure if this cited correctly  
@@ -140,6 +150,7 @@
     booktitle={Genetic Programming},
     pages={197--212},
     year={2019},
+    doi={10.1007/978-3-030-16670-0_13},
 }
 
 @article{pareto-optimality,
@@ -151,6 +162,7 @@
     article-number={37},
     year={2022},
     publisher={MDPI}
+    doi={10.3390/computation10030037}
 }
 
 @inproceedings{EvoNAS-lemonade-multi-objective,
@@ -158,6 +170,7 @@
     title={Efficient Multi-Objective Neural Architecture Search via Lamarckian Evolution},
     booktitle={International Conference on Learning Representations},
     year={2019},
+    doi={10.48550/arXiv.1804.09081},
 }
 
 @inproceedings{EvoNAS-human-sketches,
@@ -168,6 +181,7 @@
     number={},
     pages={1-5},
     year={2017},
[…]
