### 2024-11-15 44d2b41
Added doi to paper references
