+++ paper.bib
@@ -161,8 +161,8 @@
     number={3},
     article-number={37},
     year={2022},
-    publisher={MDPI}
-    doi={10.3390/computation10030037}
+    publisher={MDPI},
+    doi={10.3390/computation10030037},
 }
 
 @inproceedings{EvoNAS-lemonade-multi-objective,
@@ -298,5 +298,5 @@
     pages={436--444},
     year={2015},
     publisher={Nature Publishing Group UK London},
-    doi={10.1038/nature14539}.
+    doi={10.1038/nature14539},
 }
