### 2025-01-03 #7480 @jbytecode
@crvernon - sure, gladly!
