### 2025-01-03 #7480 @jbytecode
@leadang42 - Hi, I'm the handling editor for this submission. First, thank you for submitting your work to JOSS. The initial step in the review process is to assign reviewers for your submission. Do you have any suggestions for potential reviewers? You can refer to the authors of the latest papers suggested by our editorial bot or use the list of people provided at the top of this thread (For example diningphil is suggested as their paper is found similar to yours, automatically, possibly by using a similarity measure).

Please provide the email addresses or GitHub handles of suitable reviewers. Kindly avoid using the @ symbol to prevent unwanted notifications. Thank you in advance for your help!
