### 2025-01-16 #7480 @leadang42
Hello @jbytecode, I apologize for the delayed response. Below are the GitHub handles of some suitable reviewers:

**Genetic Algorithm Focus:**
- `igiannantoni`
- `professoralkmin`
- `sea-shunned`

**Dashboard Focus:**
- `dheshanm`
- `B3J4y`
- `priyankagagneja`

**Reviewers of Similar JOSS Papers:**
- `idoby`
- `sepandhaghighi`
