### 2025-01-26 #7480 @jbytecode
👋👋👋 Dear @B3J4y  👋👋👋

JOSS publishes articles about open source research software.

The submission I'd like you to review is titled:

**[[PRE REVIEW]: EvoVis: Dashboard for visualizing Evolutionary Neural Architecture Search algorithms](https://github.com/openjournals/joss-reviews/issues/7480)**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/7480).

Thank you in advance!
