### 2025-01-27 #7480 @B3J4y
Hi,

Yes, I am willing to review.

Best wishes,
Jan

Am 26. Januar 2025 20:32:50 MEZ schrieb Mehmet Hakan Satman ***@***.***>:
>👋👋👋 Dear @B3J4y  👋👋👋
>
>Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
>JOSS publishes articles about open source research software.
>
>The submission I'd like you to review is titled:
>
>**[[PRE REVIEW]: EvoVis: Dashboard for visualizing Evolutionary Neural Architecture Search algorithms](https://github.com/openjournals/joss-reviews/issues/7480)**
>
>You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/7480).
>
>The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
>This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be 23 check items for each single reviewer.
>
>Thank you in advance!
>
>--
>Reply to this email directly or view it on GitHub:
>https://github.com/openjournals/joss-reviews/issues/7480#issuecomment-2614561875
>You are receiving this because you were mentioned.
>
>Message ID: ***@***.***>
