### 2025-03-10 #7728 @dheshanm
Apologies for the delay—I was traveling and didn’t have the opportunity to review the repository until recently.

I am nearly finished with my checklist and wanted to share some initial feedback:

- The data specification is particularly well done, and the included examples make understanding the requirements clear.

### Minor UI Suggestions:
- Adding tooltips to the header icons would imporve usability. I had to click through them to understant the icons used.
- Changing the color of an element on hover, particularly to a bright primary color, typically suggests that clicking on it will lead to navigation. However, the hyperparameter icons on the homepage do not function this way; instead, hovering over them merely triggers tooltips, which could potentially confuse users.

### Other pointers:
- I tested the dashboard using the provided sample run data. In almost all of the times I used the Dashboard, the 'Family Tree' section seemed to go unresponsive, and I had to refresh/reload the app to continue.
- The Browser Developer console for this app has quite a few warnings and errors - none of which affect functionality, as far as I have tested - But would be a good idea to address them.
  - Some examples include:
    - `Warning: Encountered two children with the same key, `metric-card`. Keys should be unique so that components maintain their identity across updates. Non-unique keys may cause children to be duplicated and/or omitted — the behavior is unsupported and could change in a future version.`
    - This can be fixed by ensuring unique element IDs are used when constructing the DOM element
- While the codebase has a good amount of documentation, following a consistent style (like PEP 8) would make it easier for contributors to work with the codebase. Consider running a linter to enforce a consistent coding style.

I have not opened an issue based on the above suggestions, as they are not acceptance blockers.

Thank you for your work on EvoVis! I look forward to completing my review shortly.
