### 2025-03-11 #7728 @B3J4y
@jbytecode I have a question about the point "Contribution and authorship: Has the submitting author (@leadang42) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?"

- In the repository, the main contribution to the software was made by rgroh1996, who is the second author. It appears to be a new repository where the code was published in a single commit. Therefore, I cannot determine who contributed to which part. In the current version, I am unable to verify the contributions and authorship. How should I proceed?
