### 2025-03-14 #7728 @leadang42
@B3J4y @jbytecode I made the major contribution to the software. We transferred the repository that I initially worked on under my private account while I was in the same lab as rgroh1996. Since I am now at another university, rgroh1996 has taken over the repository so the lab has full control of it. Here is the original repository: https://github.com/leadang42/EvoVis.git (It is a public repository.)

@dheshanm Thank you for your feedback! One question about the toggling for metric information: Should I keep the metric explanations but just disable the color switch from light to purple?
