### 2025-03-17 #7728 @dheshanm
> @dheshanm Thank you for your feedback! One question about the toggling for metric information: Should I keep the metric explanations but just disable the color switch from light to purple?

Yep! That would definitely help. Definely keep the explanations!
