### 2025-04-02 #7728 @dheshanm
I have completed my checklist. My only holdout is the absence of any defined tests. While the dashboard uses Dash to declaratively define HTML, the underlying logic in files such as [src/dataval.py ](https://github.com/ankilab/EvoVis/blob/main/src/dataval.py)and [src/evolution.py ](https://github.com/ankilab/EvoVis/blob/main/src/evolution.py)would benefit from the addition of automated tests.
