### 2025-04-02 #7728 @jbytecode
@dheshanm  - Fair. The logic should have tests. Addition to this, GUI events can also be tested, e.g. using Selenium. Thank you.

@leadang42 - Please consider our reviewers' suggestions and update your status. Thank you in advance.
