### 2025-04-02 #7728 @leadang42
Thank you for all your feedback, I will implement the improvement suggestions (UI enhancements, automated tests).

@jbytecode - Is this the final feedback on the code for now, or what other step is missing in the current phase of the review process?
