### 2025-04-03 #7728 @jbytecode
@leadang42 - Testing GUI with Selenium is just a suggestion of me, not really mandatory, it is a feature you would consider implementing in the future.
