### 2025-04-04 ankilab/EvoVis#1 @B3J4y
Mention in README the original repository
Thank you for clarifying the origin of this repository. It would also be helpful to mention this in the repository itself.

From https://github.com/openjournals/joss-reviews/issues/7728
