### 2025-04-04 ankilab/EvoVis#2 @B3J4y
First Run: ObsoleteAttributeException
After the setup, the run didn't work in Python 3.11. The exception is:

dash.exceptions.ObsoleteAttributeException: app.run_server has been replaced by app.run

The suggestion is to modify in app.py L71 app.run_server to app.run.

For review https://github.com/openjournals/joss-reviews/issues/7728
