### 2025-04-04 ankilab/EvoVis#3 @B3J4y
Runtime throws exceptions: module 'dash_mantine_components' has no attribute 'Col'
I use Windows. After letting the software run, it throws the following problems:

```
Traceback (most recent call last):
  File ".\EvoVis\.venv1\Lib\site-packages\flask\app.py", line 1473, in wsgi_app
    response = self.full_dispatch_request()
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".\EvoVis\.venv1\Lib\site-packages\flask\app.py", line 882, in full_dispatch_request
    rv = self.handle_user_exception(e)
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".\.venv1\Lib\site-packages\flask\app.py", line 878, in full_dispatch_request
    rv = self.preprocess_request()
         ^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".\.venv1\Lib\site-packages\flask\app.py", line 1253, in preprocess_request
    rv = self.ensure_sync(before_func)()
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".\.venv1\Lib\site-packages\dash\dash.py", line 2335, in router
    [
  File ".\.venv1\Lib\site-packages\dash\dash.py", line 2336, in <listcomp>
    page["layout"]() if callable(page["layout"]) else page["layout"]
    ^^^^^^^^^^^^^^^^
  File ".\src\pages\hyperparameters_page.py", line 116, in hyperparameter_layout
    dmc.Col(parameter_overview_header(), span='auto', className='hyperparameters-header'),
    ^^^^^^^
AttributeError: module 'dash_mantine_components' has no attribute 'Col'

```
For review https://github.com/openjournals/joss-reviews/issues/7728
