### 2025-04-27 12bef6a
Metric Cards UI Enhancement
