### 2025-04-27 4eb0ad7
Added tooltips to nav bar
