+++ src/app.py
@@ -1,6 +1,7 @@
 import dash
 from dash import html
 from dash_iconify import DashIconify
+import dash_mantine_components as dmc
 
 
 ############################################################
@@ -16,31 +17,61 @@ from dash_iconify import DashIconify
 ### LAYOUT COMPONENTS
 def navbar():
     """
-    Generate the navigation bar of EvoVis.
-
+    Generate the navigation bar of EvoVis with tooltips for better usability.
+    
     Returns:
-        dash.html.Div: Navigation bar containing links to different pages.
+        dash.html.Div: Navigation bar containing links to different pages with tooltips.
     """
+    # Header descriptions and configurations
+    nav_items = [
+        {"desc": "Hyperparameter Overview", "icon": "streamline:input-box-solid", "id": "hyperparameter-link", 
+         "path": dash.page_registry['pages.hyperparameters_page']['relative_path']},
+        {"desc": "Gene Pool", "icon": "jam:dna", "id": "genepool-link", 
+         "path": dash.page_registry['pages.genepool_page']['relative_path']},
+        {"desc": "Family Tree", "icon": "mdi:graph", "id": "family-tree-link", 
+         "path": dash.page_registry['pages.family_tree_page']['relative_path']},
+        {"desc": "Run Results", "icon": "entypo:bar-graph", "id": "results-link", 
+         "path": dash.page_registry['pages.run_results_page']['relative_path']}
+    ]
+    
+    # Create nav links with tooltips
+    nav_links = []
+    for item in nav_items:
+        nav_links.append(
+            dmc.Tooltip(
+                label=item["desc"],
+                position="bottom",
+                offset=5,
+                radius=15,
+                transition="pop-top-left",
+                color="#6173E9",
+                multiline=False,
+                width="auto",
+                children=[
+                    html.A(
+                        html.Button(
+                            children=DashIconify(icon=item["icon"], height=25, width=25, color="#000000"), 
+                            className="circle-btn", 
+                            id=item["id"]
+                        ), 
+                        href=item["path"],
+                        style={"display": "inline-block"}
+                    )
+                ]
+            )
+        )
+    
     return html.Div(
-    [
-        html.Div(
-            [   
-                html.A(children=html.Img(src="assets/media/evonas-logo.png", height="50px"), href=dash.page_registry['pages.hyperparameters_page']['relative_path']),
-            ],
-            id="navrun"
-        ),
-        html.Div(
-            [
-                html.A(html.Button(children=DashIconify(icon="streamline:input-box-solid", height=25, width=25, color="#000000"), className="circle-btn", id="hyperparameter-link"), href=dash.page_registry['pages.hyperparameters_page']['relative_path']),
-                html.A(html.Button(children=DashIconify(icon="jam:dna", height=25, width=25, color="#000000"), className="circle-btn", id="genepool-link"), href=dash.page_registry['pages.genepool_page']['relative_path']),
-                html.A(html.Button(children=DashIconify(icon="mdi:graph", height=25, width=25, color="#000000"), className="circle-btn", id="family-tree-link"), href=dash.page_registry['pages.family_tree_page']['relative_path']),
-                html.A(html.Button(children=DashIconify(icon="entypo:bar-graph", height=25, width=25,color="#000000"), className="circle-btn", id="results-link"), href=dash.page_registry['pages.run_results_page']['relative_path']), 
-            ],
-            id="navlinks",
-        )    
-    ],
-    id="navbar",
-)
+        [
+            html.Div(
+                [html.A(children=html.Img(src="assets/media/evonas-logo.png", height="50px"), 
[…]
