### 2025-04-27 7e9b535
Safeguard for app run function
