### 2025-04-27 ankilab/EvoVis#2 @leadang42
I've just created a virtual environment with Python 3.11.3 (the version I developed the app with and specified in the README) to test the issue. I didn't encounter the same error. However, to ensure compatibility across different Dash versions, I've implemented a solution that handles both API versions:

```python
if __name__ == "__main__":
    try:
        # Try newer method first
        app.run(host="0.0.0.0", port=8050, debug=True)
    except AttributeError:
        # Fall back to the older method
        app.run_server(host="0.0.0.0", port=8050, debug=True)
```

This implementation will try the newer `app.run()` method first, and if that fails, it will fall back to the older `app.run_server()` method. This should make the application more robust across different Dash versions and development environments.
