### 2025-04-27 ankilab/EvoVis#3 @leadang42
Have you checked that you're using the correct version of Dash Mantine Components (0.12.0 specified in the requirements.txt)? I've researched, and it appears that in version 0.14.0, the `Col` component was renamed to `GridCol`. Given that the error you're having is about the Mantine package, I think it's unlikely to be a Windows-related issue.
