+++ src/pages/genepool_page.py
@@ -16,7 +16,7 @@ run = os.getenv("RUN_RESULTS_PATH")
 
 
 ### REGISTER DASH APP
-dash.register_page(__name__, path='/genepool')
+dash.register_page(__name__, path="/genepool")
 
 
 ### GENE POOL PAGE COMPONENTS
@@ -32,94 +32,89 @@ def cytoscape_stylesheet(groups):
     """
     cytoscape_stylesheet = [
         {
-            'selector': 'node',
-            'style': {
-                'background-color': '#6173E9',
-                'content': 'data(label)',
-                'width':'50px',
-                'height':'50px',
-            }
+            "selector": "node",
+            "style": {
+                "background-color": "#6173E9",
+                "content": "data(label)",
+                "width": "50px",
+                "height": "50px",
+            },
+        },
+        {"selector": "edge", "style": {"line-color": "#6173E9", "width": "2px"}},
+        {
+            "selector": "label",
+            "style": {
+                "font-family": "sans-serif",
+                "color": "#FFFFFF",
+                "font-size": "12px",
+                "font-weight": "500",
+                "text-valign": "center",
+            },
         },
         {
-            'selector': 'edge',
-            'style': {
-                'line-color': '#6173E9',
-                'width': '2px'
-            }
+            "selector": '[id = "Start"]',
+            "style": {
+                "background-color": "#FFFFFF",
+                "border-color": "#6173E9",
+                "border-width": "2px",
+                "content": "data(label)",
+                "color": "#000000",  # Font color
+            },
         },
         {
-            'selector': 'label',
-            'style': {
-                'font-family': 'sans-serif',
-                'color': '#FFFFFF',
-                'font-size': '12px',
-                'font-weight': '500',
-                'text-valign': 'center',
-            }
+            "selector": '[id = "end"]',
+            "style": {
+                "background-color": "#FFFFFF",
+                "border-color": "#6173E9",
+                "border-width": "2px",
+                "content": "data(label)",
+                "color": "#000000",  # Font color
+            },
         },
         {
-            'selector': '[id = "Start"]',
-            'style': {
-                'background-color': '#FFFFFF',
-                'border-color': '#6173E9',
-                'border-width': '2px',
-                'content': 'data(label)',
-                'color': '#000000', # Font color
-            }
-        },
-        {
-            'selector': '[id = "end"]',
-            'style': {
-                'background-color': '#FFFFFF',
-                'border-color': '#6173E9',
-                'border-width': '2px',
-                'content': 'data(label)',
-                'color': '#000000', # Font color
-            }
-        },
-        {
-            'selector': '[id = "class-connect"]',
-            'style': {
-                'width': '2px',
-                'line-color': '#6173E9',
-            }
+            "selector": '[id = "class-connect"]',
+            "style": {
+                "width": "2px",
+                "line-color": "#6173E9",
+            },
         },
     ]
 
     # Style for groupe nodes
     for group in groups:
-    
+
         group_node_style = {
-            'selector': f'[id = "{group}"]',
-            'style': {
-                #Box
-                'background-fill': 'radial-gradient', 
-                'background-gradient-stop-colors': '#EFEFEF #6173E9',
-                'background-gradient-stop-positions': '0% 50%',
-                'background-opacity': '0.5',
-                'shape': 'round-rectangle',
[…]
