+++ src/pages/hyperparameters_page.py
@@ -13,7 +13,7 @@ run = os.getenv("RUN_RESULTS_PATH")
 
 
 ### REGISTER DASH APP
-dash.register_page(__name__, path='/')
+dash.register_page(__name__, path="/")
 
 
 ### HYPERPARAMETER PAGE COMPONENTS
@@ -24,16 +24,16 @@ def grouped_metriccards():
     Each group is displayed as a header along with the corresponding hyperparameter metric cards.
 
     Returns:
-        list: A list of HTML div elements representing the hyperparameter groups, 
+        list: A list of HTML div elements representing the hyperparameter groups,
               where each group includes the group name as the header and the associated hyperparameter metric cards.
     """
-    
+
     # Retrieve hyperparameters for the specified run
     hps = get_hyperparameters(run)
-    
+
     # Initialize dictionary to store hyperparameters grouped by their respective groups
     groups = {}
-    
+
     # Initialize list to store hyperparameters without a group
     other_params = []
 
@@ -42,18 +42,28 @@ def grouped_metriccards():
         value = hp.get("value", None)
         display = hp.get("display", True)
 
-        if value is not None and display: 
+        if value is not None and display:
             # Extract values of hyperparameter and handle missing keys
             unit = hp.get("unit", None)
-            icon = hp.get("icon", "icon-park:expand-text-input") or "icon-park:expand-text-input"
+            icon = (
+                hp.get("icon", "icon-park:expand-text-input")
+                or "icon-park:expand-text-input"
+            )
             displayname = hp.get("displayname", hp_key) or hp_key
-            group = hp.get("group", None) 
+            group = hp.get("group", None)
             description = f"{displayname}: {hp.get('description', displayname)}"
-            metrictype = displayname or hp_key.replace("_", " ").capitalize() 
+            metrictype = displayname or hp_key.replace("_", " ").capitalize()
             metric = str(hp.get("value", "")).replace("\n", "")
-            
+
             # Create metric card div component for hyperparameter visual
-            mc = parameter_card(metrictype=metrictype, metric=metric, icon=icon, unit=unit, description=description, max_width=280)
+            mc = parameter_card(
+                metrictype=metrictype,
+                metric=metric,
+                icon=icon,
+                unit=unit,
+                description=description,
+                max_width=280,
+            )
 
             # Add the metric card to its group or to the list of hyperparameters without a group
             if group is None:
@@ -62,21 +72,21 @@ def grouped_metriccards():
                 groups[group] = [mc]
             else:
                 groups[group].append(mc)
-    
 
     # Add hyperparameters without a group to the end
-    if len(other_params) != 0:   
+    if len(other_params) != 0:
         groups["Other parameters"] = other_params
 
-    # Convert the group dictionaries to list of divs  
+    # Convert the group dictionaries to list of divs
     group_divs = []
 
     # Create group div with group heading and metric cards for each group
-    for group_key, group in groups.items():   
+    for group_key, group in groups.items():
         group_div = html.Div([dot_heading(group_key)] + group, className="metrics")
         group_divs.append(group_div)
 
     return group_divs
+
 
 def parameter_overview_header():
     """Generate hyperparameter page header.
@@ -91,7 +101,7 @@ def parameter_overview_header():
     ]
 
 
-### HYPERPARAMETER PAGE LAYOUT 
+### HYPERPARAMETER PAGE LAYOUT
 def hyperparameter_layout():
     """
     Generates the layout for the hyperparameter page.
@@ -100,21 +110,29 @@ def hyperparameter_layout():
     Returns:
         dash_mantine_components.Grid: A Mantine Grid component representing the layout of the hyperparameter page.
     """
-    
+
     validation_result = validate_hyperparameters(run)
[…]

[…]
