+++ src/dataval.py
@@ -41,7 +41,9 @@ def json_to_dict(filepath):
         raise FileNotFoundError(f"File not found: {filepath}")
 
     except json.JSONDecodeError as e:
-        raise json.JSONDecodeError(f"Error decoding JSON data in file {filepath}: {e}")
+        raise json.JSONDecodeError(
+            f"Error decoding JSON data in file {filepath}: {str(e)}", e.doc, e.pos
+        )
 
 
 def validate_hyperparameters(run):
