+++ tests/test_evolution.py
@@ -1,0 +1,233 @@
+import pytest
+import json
+from unittest.mock import patch, mock_open
+
+from src.evolution import (
+    _json_to_dict,
+    get_generations,
+    get_hyperparameters,
+    get_meas_info,
+    get_individual_result,
+    get_individual_chromosome,
+    get_individuals,
+    get_healthy_individuals_results,
+    get_best_individuals,
+)
+
+
+@pytest.fixture
+def mock_config_data():
+    return {
+        "hyperparameters": {"population_size": 10, "mutation_rate": 0.1},
+        "results": {
+            "accuracy": {
+                "displayname": "Accuracy",
+                "unit": "%",
+                "min-boundary": 0,
+                "max-boundary": 100,
+            },
+            "loss": {"displayname": "Loss", "unit": None},
+        },
+    }
+
+
+@pytest.fixture
+def mock_results_data():
+    return {"accuracy": 85.5, "loss": 0.12, "fitness": 0.75}
+
+
+@pytest.fixture
+def mock_chromosome_data():
+    return [
+        {"layer": "STFT_2D", "group": "Feature Extraction 2D"},
+        {"layer": "MAG_2D", "group": "Feature Extraction 2D"},
+    ]
+
+
+@pytest.fixture
+def mock_directory_structure():
+    directories = ["Generation_1", "Generation_2", "Generation_3"]
+
+    with patch("os.path.exists", return_value=True), patch(
+        "os.listdir", return_value=directories
+    ), patch("os.path.isdir", return_value=True):
+        yield directories
+
+
+# Tests for _json_to_dict
+def test_json_to_dict_success():
+    mock_data = '{"key": "value"}'
+    with patch("builtins.open", mock_open(read_data=mock_data)):
+        result = _json_to_dict("dummy_path.json")
+        assert result == {"key": "value"}
+
+
+def test_json_to_dict_file_not_found():
+    with patch("builtins.open", side_effect=FileNotFoundError):
+        with pytest.raises(FileNotFoundError):
+            _json_to_dict("non_existent_file.json")
+
+
+def test_json_to_dict_invalid_json():
+    mock_data = "{invalid json}"
+    with patch("builtins.open", mock_open(read_data=mock_data)):
+        with pytest.raises(json.JSONDecodeError):
+            _json_to_dict("invalid_json.json")
+
+
+# Tests for get_generations
+def test_get_generations(mock_directory_structure):
+    with patch("src.evolution._get_individuals_of_generation") as mock_get_individuals:
+        mock_get_individuals.return_value = {"individual_1": {"fitness": 0.8}}
+
+        with patch("os.path.exists", return_value=True):
+            result = get_generations("test_run")
+            assert result == ["Generation_1", "Generation_2", "Generation_3"]
+
+
+def test_get_generations_as_int(mock_directory_structure):
+    with patch("src.evolution._get_individuals_of_generation") as mock_get_individuals:
+        mock_get_individuals.return_value = {"individual_1": {"fitness": 0.5}}
+
+        with patch("os.path.exists", return_value=True):
+            result = get_generations("test_run", as_int=True)
+            assert result == [1, 2, 3]
+
+
+# Tests for get_hyperparameters and get_meas_info
+def test_get_hyperparameters(mock_config_data):
+    with patch("src.evolution._get_configurations", return_value=mock_config_data):
+        result = get_hyperparameters("test_run")
+        assert result == {"population_size": 10, "mutation_rate": 0.1}
+
+
+def test_get_meas_info(mock_config_data):
+    with patch("src.evolution._get_configurations", return_value=mock_config_data):
+        result = get_meas_info("test_run")
+
+        # Check that default values are applied
[…]
