+++ conftest.py
@@ -1,0 +1,6 @@
+import sys
+import os
+from pathlib import Path
+
+# Add the project root directory to sys.path
+sys.path.append(str(Path(__file__).parent))
