+++ tests/test_dataval.py
@@ -16,9 +16,9 @@ from src.dataval import (
 )
 
 
-# Test fixtures
 @pytest.fixture
 def mock_config_data():
+    """Fixture providing valid mock configuration data with hyperparameters and results."""
     return {
         "hyperparameters": {
             "population_size": {"value": 10},
@@ -33,6 +33,7 @@ def mock_config_data():
 
 @pytest.fixture
 def mock_invalid_config_data():
+    """Fixture providing invalid configuration data with structural problems."""
     return {
         "hyperparameters": {
             "population_size": 10,  # Missing dictionary structure
@@ -43,6 +44,7 @@ def mock_invalid_config_data():
 
 @pytest.fixture
 def mock_search_space_data():
+    """Fixture providing valid search space data with gene pools and rule sets."""
     return {
         "gene_pool": {
             "Feature Extraction": [
@@ -59,6 +61,7 @@ def mock_search_space_data():
 
 @pytest.fixture
 def mock_invalid_search_space_data():
+    """Fixture providing invalid search space data with missing keys and incorrect types."""
     return {
         "gene_pool": {
             "Feature Extraction": [
@@ -74,14 +77,15 @@ def mock_invalid_search_space_data():
 
 @pytest.fixture
 def mock_crossover_parents_data():
+    """Fixture providing formatted crossover parents data string."""
     # Fixed format to match what the function expects
     return (
         "Generation: 1,Parent_1: (ind1, 2),Parent_2: (ind2, 3),New_Individual: new_ind"
     )
 
 
-# Tests for json_to_dict
 def test_json_to_dict_success():
+    """Test that json_to_dict correctly parses valid JSON files."""
     mock_data = '{"key": "value"}'
     with patch("builtins.open", mock_open(read_data=mock_data)):
         result = json_to_dict("dummy_path.json")
@@ -89,20 +93,22 @@ def test_json_to_dict_success():
 
 
 def test_json_to_dict_file_not_found():
+    """Test that json_to_dict raises FileNotFoundError when file doesn't exist."""
     with patch("builtins.open", side_effect=FileNotFoundError):
         with pytest.raises(FileNotFoundError):
             json_to_dict("non_existent_file.json")
 
 
 def test_json_to_dict_invalid_json():
+    """Test that json_to_dict raises JSONDecodeError for invalid JSON."""
     mock_data = "{invalid json}"
     with patch("builtins.open", mock_open(read_data=mock_data)):
         with pytest.raises(json.JSONDecodeError):
             json_to_dict("invalid_json.json")
 
 
-# Tests for validate_hyperparameters
 def test_validate_hyperparameters_success(mock_config_data):
+    """Test that validate_hyperparameters returns empty string for valid hyperparameters."""
     with patch("os.path.exists", return_value=True), patch(
         "src.dataval.json_to_dict", return_value=mock_config_data
     ):
@@ -111,12 +117,14 @@ def test_validate_hyperparameters_success(mock_config_data):
 
 
 def test_validate_hyperparameters_file_not_found():
+    """Test that validate_hyperparameters detects missing configuration file."""
     with patch("os.path.exists", return_value=False):
         result = validate_hyperparameters("test_run")
         assert "not found" in result
 
 
 def test_validate_hyperparameters_invalid_structure(mock_invalid_config_data):
+    """Test that validate_hyperparameters detects invalid hyperparameter structure."""
     with patch("os.path.exists", return_value=True), patch(
         "src.dataval.json_to_dict", return_value=mock_invalid_config_data
     ):
@@ -124,8 +132,8 @@ def test_validate_hyperparameters_invalid_structure(mock_invalid_config_data):
[…]
