+++ tests/test_genepool.py
@@ -2,7 +2,6 @@ import pytest
 import json
 from unittest.mock import patch, mock_open, MagicMock
 
-# Import the functions to test
 from src.genepool import (
     _json_to_dict,
     _get_search_space,
@@ -12,9 +11,9 @@ from src.genepool import (
 )
 
 
-# Test data fixtures
 @pytest.fixture
 def mock_json_data():
+    """Fixture providing mock search space data with gene pool, rule sets, and group rules."""
     return {
         "gene_pool": {
             "Feature Extraction 2D": [
@@ -43,12 +42,12 @@ def mock_json_data():
 
 @pytest.fixture
 def mock_graph():
+    """Fixture providing a mock layer graph with connections between layers."""
     return {"Start": ["STFT_2D"], "STFT_2D": ["MAG_2D"], "MAG_2D": ["C_2D", "DC_2D"]}
 
 
-# Tests for _json_to_dict
 def test_json_to_dict_success():
-    # Test successful JSON file reading
+    """Test that _json_to_dict correctly parses valid JSON files."""
     mock_data = '{"key": "value"}'
     with patch("builtins.open", mock_open(read_data=mock_data)):
         result = _json_to_dict("dummy_path.json")
@@ -56,26 +55,23 @@ def test_json_to_dict_success():
 
 
 def test_json_to_dict_file_not_found():
-    # Test FileNotFoundError handling
+    """Test that _json_to_dict raises FileNotFoundError when file doesn't exist."""
     with patch("builtins.open", side_effect=FileNotFoundError):
         with pytest.raises(FileNotFoundError):
             _json_to_dict("non_existent_file.json")
 
 
 def test_json_to_dict_invalid_json():
-    # Test invalid JSON handling
+    """Test that _json_to_dict raises JSONDecodeError for invalid JSON."""
     mock_data = "{invalid json}"
 
-    # We need to patch the json.JSONDecodeError that's raised internally
-    # Rather than expecting the custom message
     with patch("builtins.open", mock_open(read_data=mock_data)):
         with pytest.raises(json.JSONDecodeError):
             _json_to_dict("invalid_json.json")
 
 
-# Tests for _get_search_space
 def test_get_search_space():
-    # Test that _get_search_space calls _json_to_dict with the correct path
+    """Test that _get_search_space loads search space from the correct file path."""
     with patch("src.genepool._json_to_dict") as mock_json_to_dict:
         mock_json_to_dict.return_value = {"key": "value"}
         result = _get_search_space("test_run")
@@ -84,9 +80,8 @@ def test_get_search_space():
         assert result == {"key": "value"}
 
 
-# Tests for _get_layer_graph
 def test_get_layer_graph_without_group_connections(mock_json_data):
-    # Test layer graph generation without group connections
+    """Test that _get_layer_graph builds correct layer connections without group connections."""
     with patch("src.genepool._get_search_space", return_value=mock_json_data):
         result = _get_layer_graph("test_run", group_connections=False)
 
@@ -99,8 +94,7 @@ def test_get_layer_graph_without_group_connections(mock_json_data):
 
 
 def test_get_layer_graph_with_group_connections(mock_json_data):
-    # Test layer graph generation with group connections
-    # This is more complex because it involves _get_groups too
+    """Test that _get_layer_graph includes group connections when requested."""
     with patch("src.genepool._get_search_space", return_value=mock_json_data):
         with patch("src.genepool._get_groups") as mock_get_groups:
             mock_get_groups.return_value = {
@@ -110,24 +104,21 @@ def test_get_layer_graph_with_group_connections(mock_json_data):
 
             result = _get_layer_graph("test_run", group_connections=True)
 
-            # This should include both direct connections and group-based connections
[…]

[…]
