### 2025-04-28 #7728 @leadang42
I have reviewed all open issues and improvement suggestions, and addressed most of them:

- **Metric Cards UI Enhancement**: Disabled the color switch from light to purple on hover as they are not clickable buttons.
- **Navigation Bar UI Enhancement**: Added tooltips to navigation buttons to improve usability without requiring clicks.
- **Repository Disclaimer in README**: Clarified and mentioned both the old and new repositories (`ankilab/EvoVis` and `leadang42/EvoVis`).
- **Unit Tests**: Added unit tests for the core logic scripts:`test_genepool.py`, `test_evolution.py`, and `test_datival.py` now test `genepool.py`, `evolution.py`, and `datival.py`. (Note: The remaining scripts are primarily UI generators and do not include complex data-processing logic, so no unit tests were added for them.)
- **Code Linting**: Ran the `black` code formatter to enforce PEP 8 compliance.

---

Regarding the **Family Tree responsiveness issue**:
After investigation, I believe there are two main situations that can make it appear unresponsive:

1. **Expanding a node for an extinct individual**:  No further nodes are added, as there are no descendants.

2. **Expanding a very large tree**:  The tree *does* continue to expand, but it takes a long time due to the recursive nature of both the underlying data structure and the rendering logic.

**Proposed improvement**:
A simple way to address this would be to add a **progress bar** during the family tree generation.
This would help make it clear:
- When no new nodes appear because the individual is extinct.
- When the expansion is simply taking longer due to tree size (the progress bar would still be active).

---

Let me know if you'd like me to prioritize adding the progress bar next! @jbytecode
