### 2025-04-29 #7728 @dheshanm
Thanks for the updates @leadang42. I believe these UI enhancements definitively improve user's experience with the dashboard.

About the Family Tree responsiveness, I agree that is could be due to unraveling a large recursive tree. If adding a progress bar to the GUI takes too much effort, having logs indicating the progress printed to the console could be a good alternative, that might be easier to implement.

@jbytecode: With the unit tests implemented, and code reformatted to comply with PEP8, all items on my checklist are checked off.
