### 2025-04-30 5df6415
Added loading prop for family tree generation
