+++ src/pages/family_tree_page.py
@@ -1,5 +1,5 @@
 import dash
-from dash import html, callback, Input, Output, dcc
+from dash import html, callback, Input, Output, State, dcc
 import dash_cytoscape as cyto
 import dash_mantine_components as dmc
 from dash_iconify import DashIconify
@@ -97,15 +97,28 @@ MARKS_STYLE = {
 def family_tree_cytsocape():
     """
     Generates a Dash Cytoscape component representing the family tree.
+    Wrapped in a LoadingOverlay component.
 
     Returns:
-        dash_cytoscape.Cytoscape: Dash Cytoscape component representing the family tree.
+        dash_mantine_components.LoadingOverlay: Dash Mantine LoadingOverlay component wrapping the Cytoscape component.
     """
-    return cyto.Cytoscape(
-        id="cytoscape-family-tree",
-        className="wrapper",
-        style={"height": "450px", "max-width": "100%"},
-        stylesheet=CYTOSCAPE_STYLE,
+    return html.Div(
+        [
+            dmc.LoadingOverlay(
+                cyto.Cytoscape(
+                    id="cytoscape-family-tree",
+                    className="wrapper",
+                    style={"height": "430px", "max-width": "100%"},
+                    stylesheet=CYTOSCAPE_STYLE,
+                ),
+                id="family-tree-loading",
+                #loaderProps={"color": "#262626", "size": "xl", "variant": "dots"},
+                overlayOpacity=0,  # Transparent background
+                overlayColor="transparent",  # Transparent color
+                overlayBlur=2,  # Add blur effect (values typically range from 1-5)
+            )
+        ],
+        id="cytoscape-container"
     )
 
 
@@ -121,25 +134,28 @@ def generation_slider():
         run, generations_int[round(len(generations_int) / 3)]
     )
 
-    return dcc.RangeSlider(
-        min(generations_int),
-        max(generations_int),
-        1,
-        marks={
-            min(generations_int): {
-                "label": f"Generation_{min(generations_int)}",
-                "style": MARKS_STYLE,
+    return html.Div(
+        dcc.RangeSlider(
+            min(generations_int),
+            max(generations_int),
+            1,
+            marks={
+                min(generations_int): {
+                    "label": f"Generation_{min(generations_int)}",
+                    "style": MARKS_STYLE,
+                },
+                max(generations_int): {
+                    "label": f"Generation_{max(generations_int)}",
+                    "style": MARKS_STYLE,
+                },
             },
-            max(generations_int): {
-                "label": f"Generation_{max(generations_int)}",
-                "style": MARKS_STYLE,
-            },
-        },
-        # pushable=1,
-        allowCross=False,
-        id="gen-range-slider",
-        tooltip={"placement": "top", "always_visible": True},
-        value=[random_generation - 4, random_generation, random_generation + 1],
+            # pushable=1,
+            allowCross=False,
+            id="gen-range-slider",
+            tooltip={"placement": "top", "always_visible": True},
+            value=[random_generation - 4, random_generation, random_generation + 1],
+        ),
+        style={"margin-top": "12px"}  # Add padding/margin at the top of the slider so not blurred by family tree laoding
     )
 
 
@@ -195,17 +211,40 @@ def set_individuals_select(gen_range):
 
 
 @callback(
-    Output("cytoscape-family-tree", "elements"),
-    Output("cytoscape-family-tree", "layout"),
-    Output("cytoscape-family-tree", "stylesheet"),
-    Input("gen-range-slider", "value"),
-    Input("ind-select", "value"),
-    Input("cytoscape-family-tree", "tapNodeData"),
[…]
