### 2025-04-30 #7728 @leadang42
@dheshanm - Thank you for the feedback.

I've just added the loading state of the family tree: When the recursive tree is generated, the old tree (1) gets blurred and has (2) loading points to indicate the generation of the new tree.
