### 2025-04-30 8be5274
Improved stylig of loading props for family tree
