### 2025-05-31 #7728 @leadang42
@jbytecode – I’ve responded to the issue and explained that the error `module 'dash_mantine_components' has no attribute 'Col'` is due to changes in the latest Dash Mantine releases, which modified parts of their API. However, I’ve double-checked, and when running the dashboard with the Mantine version specified in `requirements.txt`, it works correctly. Can I close the issue?
