### 2025-06-19 ankilab/EvoVis#3 @B3J4y
Yes, you are right. My libraries were in a current version. I have set them now to the version which was written in the requirements.txt and everything is working. I don't know why it was not initially done
