### 2025-06-19 ankilab/EvoVis#4 @B3J4y
How to contribute
Maybe I just missed it, but I haven't seen any contribution statement. In the review, there is the question: "Community guidelines: Are there clear guidelines for third parties wishing to: 1. Contribute to the software, 2. Report issues or problems with the software, 3. Seek support?"

Support information is provided at the bottom. However, if someone wants to contribute, they won't know what is appropriate. Perhaps you could use a generator to create a CONTRIBUTING.md file that specifies exactly what you want without having to formulate it yourself.

For review https://github.com/openjournals/joss-reviews/issues/7728
