### 2025-07-06 2ac4b4c
Defined ways to contribute in CONTRIBUTING.md
