+++ CONTRIBUTING.md
@@ -1,0 +1,50 @@
+# 🧩 How to Contribute
+
+Thank you for your interest in contributing! We welcome contributions from the community to improve and expand this project. This document outlines how you can get involved.
+
+## 1. Reporting Issues
+
+If you encounter a bug, have a question, or want to suggest an enhancement:
+
+- Check the existing issues to see if your concern is already addressed.
+- Open a new issue and use the provided template, if available.
+- Be as specific as possible, including steps to reproduce the issue and system details (OS, version, etc.).
+
+## 2. Submitting Pull Requests
+
+To submit a code or documentation change:
+
+1. Fork the repository.
+2. Create a new branch:
+
+   ```bash
+   git checkout -b your-feature-name
+   ```
+
+3. Make your changes.
+4. Run any tests or checks, if applicable.
+5. Commit with a clear message:
+
+   ```bash
+   git commit -m "Add new feature X"
+   ```
+
+6. Push your branch and open a pull request.
+
+Please ensure:
+
+- Your code is readable, well-commented, and follows project style guidelines.
+- Any new features or changes include appropriate tests and documentation.
+- You link the pull request to a related issue if applicable (e.g., `Closes #42`).
+
+## 3. Documentation Contributions
+
+Documentation improvements are always appreciated! Feel free to submit PRs for:
+
+- Typos or grammar fixes
+- Clarifications
+- Examples or tutorials
+
+## 4. Feature Requests
+
+Before implementing a large new feature, please open an issue to discuss your idea and get feedback. This helps avoid duplicate work and ensures alignment with the project's goals.
