### 2025-07-06 ankilab/EvoVis#4 @leadang42
I have now added the CONTRIBUTING.md file to specify how externals can contribute to the project.
