### 2025-07-08 #7728 @jbytecode
May I request a status update for this review? Thank you in advance.
