### 2025-07-11 #7728 @jbytecode
@leadang42 - I have just sent a pull request. This PR fixes a minor issue of the paper. Please review the PR. If you are agree with the changes, please merge. Thank you in advance.

https://github.com/ankilab/EvoVis/pull/5
