### 2025-07-17 98ce056
Created CITATION.cff for zenodo
