### 2025-07-17 f169c62
Added CITATION.cff acoording to Zenodo
