+++ CITATION.cff
@@ -2,7 +2,7 @@ cff-version: 1.1.0
 message: "If you use this software, please cite it as below."
 title: "EvoVis: Dashboard for Visualizing Evolutionary Neural Architecture Search Algorithms"
 version: "v1.0.0"
-date-released: 2024-11-12
+date-released: 2024-07-17
 keywords:
   - Python
   - Evolutionary neural architecture search algorithm
