### 2025-07-17 #7728 @leadang42
@jbytecode - Here are the required links:

- Release Link: [Github](https://github.com/ankilab/EvoVis/releases/tag/v1.0.0)
- Archive Link: [10.5281/zenodo.16050739](https://zenodo.org/records/16050740)
- Archive Doi: 10.5281/zenodo.16050739
