### 2025-07-18 7f7727e
Update paper.bib
Updated the DOI, it was missing a 9 at the end
