+++ paper.bib
@@ -77,7 +77,7 @@
     booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference},
     pages={419--427},
     year = {2019},
-    doi={10.1145/3321707.332172},
+    doi={10.1145/3321707.3321729},
 }
 
 @inproceedings{EvoNAS-rene, 
