### 2025-07-18 0d296e8
Update paper.md fixing typos and grammar
