### 2025-07-21 #7728 @arfon
Unfortunately I don't think we can fix this. We will need to update the bibtex to point to `url = https://ricerca.uniba.it/handle/11586/380827` and remove the DOI.

That DOI isn't going to pass our regex checks.
