### 2025-07-21 eaa5b5b
Update paper.bib
Fix DOI problem with JOSS editorialbot
