+++ paper.bib
@@ -248,7 +248,7 @@
     volume={3078},
     pages={65--76},
     year={2022},
-    doi={11586/380827}
+    url={https://ricerca.uniba.it/handle/11586/380827}
 }
 
 @article{cytoscape-js,
