### 2025-07-21 #7728 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
