### 2025-07-23 #7728 @crvernon
@anki-xyz and @jbytecode - very clean submission.  Nice work!
