### 2025-07-23 #7728 @crvernon
🥳 Congratulations on your new publication @leadang42 and @anki-xyz! Many thanks to @jbytecode for editing and @B3J4y and @dheshanm for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
